Simple "up" button functionality
This uses a parent option set in each view. In the future we can expand
this to find a parent programmatically, but for now it's just a
hard-coded path that never changes.

In addition, this commit switches to material-ui 1.0.0 beta 4. There are
some troublesome breaking changes from the previous stable release, so
switching now should save us some headaches.