Skip rendering dialog contents when hidden (#4)
This adds some extra code in our paper-dialog wrapper that keeps track
of visibility. To do this, we need to handle the neon-animation-finish
event and update component state.

If the dialog is animating through other means (outside of Polymer),
this should also work correctly - the child will be skipped as long as
the dialog's "opened" property is false.