Issue #31 Include jsLingui cache during project clean command
Also change jsLingui sourceDir config back to /src/ as the current
pinned version of jsLingui does not know how to parse TypeScript.

The locale/_build dir is a cache used by jsLingui to keep track of the messages
it has extracted, whether they have changed and where they came from. When
switching branches (eg: testing pull requests) this cache gets corrupted
(caches entries for messages that no-longer exist or have never existed). This
was particularly problematic for me when testing out newer versions of jsLingui
that have the ability to parse TypeScript.

I suspect this cache is likely the main cause of the confusion we have
about the locale catalogs being inconsistently generated.