Issue #74 Use a different Action for each page
- Each page type (Questionnaire, MyPlan, ExploreAll) now has its own Action type to represent switching to that main page
- This prevents issues with stringifying the Page enum in the path when using something like `:mainPage` in the `routesMap`
- The router now now has a function that will map a Page enum into a string route. I don't see any need for any file outside the router to use this function, so I didn't export it. Any other function should be interacting with the route by triggering one of these three Actions
I found having the `setMainPage` function in `page_switcher.ts` sort of confusing. In an ideal situation we could be using the `<Link>` tags within the `.tsx` files and not even have to define these dispatches. So I moved them into the `index.ts` where they can be closer to where they are used
- Removed some other unused code
- Updated tests